+++ src/strings_and_hash/strings.go
@@ -2,7 +2,7 @@ package strings_and_hash
 
 import (
 	"fmt"
-	"go_redis/persistence"
+	"go_redis/src/persistence"
 	"time"
 )
